requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
torch>=1.10.0
torchvision>=0.11.0
pytorch-lightning>=1.5.0
torchmetrics>=0.7.0
pyyaml>=6.0
tqdm>=4.62.0
tensorboard>=2.7.0
scikit-learn>=1.0.0
onnx>=1.10.0
onnxruntime>=1.10.0 